Conserto de problemas de fluxo de tela, retornos, gerenciamento do Navigation Stack; Reorganização de blocos de código
