+++ haub/android/app/build.gradle
@@ -62,6 +62,6 @@ flutter {
 
 dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation "com.google.firebase:firebase-auth:18.0.0"
+    implementation "com.google.firebase:firebase-auth:19.3.2"
     implementation "com.google.firebase:firebase-firestore:20.0.0"
 }
