+++ haub/android/settings_aar.gradle
@@ -1,9 +1,5 @@
 include ':app'
 
-
-
-
-
 def flutterProjectRoot = rootProject.projectDir.parentFile.toPath()
 
 def plugins = new Properties()
