+++ haub/lib/view/add_question/doubt_abstract.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-
-import '../../main.dart';
+//import 'package:haub/view/home/home.dart';
 
 class MyDaubtAbstractPage extends StatelessWidget {
   MyDaubtAbstractPage({Key key}) : super(key: key);
@@ -15,10 +14,9 @@ class MyDaubtAbstractPage extends StatelessWidget {
         child: RaisedButton(
             child: Text("Finalizar"),
             onPressed: () {
-              Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                      builder: (BuildContext context) => MyHomePageWidget()));
+              Navigator.popUntil(
+                context, 
+                ModalRoute.withName('/home'));
             }),
       ),
     );
